fix #12 by always bubbling files to the action.
if the file has errored before being added to the queue,
the `upload` and `read` functions will be stubbed with a
rejected promise. files that have errored will also be
marked as such with an `error` property, which is the
error object returned by plupload.

+++ addon/system/file.js
@@ -126,12 +126,14 @@ export default Ember.Object.extend({
     @method destroy
    */
   destroy: function () {
+    if (this.isDestroyed) { return; }
     get(this, 'uploader').removeFile(get(this, 'file'));
+    this.isDestroyed = true;
   },
 
   upload: function (url, settings) {
     var uploader = get(this, 'uploader');
-    this._deferred = RSVP.defer();
+    this._deferred = RSVP.defer(`File: '${get(this, 'id')}' Upload file`);
 
     if (settings == null) {
       settings = url;
